// projects/ngx-material-extended/src/lib/.legacy/theming/theming-directives/types.ts
export declare type ElmColor =
    | 'primary'
    | 'accent'
    | 'warn'
    | 'alert'
    | 'default';

export declare type ElmColorClass =
    | 'elm-color__default'
    | 'elm-color__primary'
    | 'elm-color__accent'
    | 'elm-color__warn'
    | 'elm-color__alert'
    | 'elm-color__default--bold'
    | 'elm-color__primary--bold'
    | 'elm-color__accent--bold'
    | 'elm-color__warn--bold'
    | 'elm-color__warn--alert';

export const elmColorClasses = [
    'elm-color__default',
    'elm-color__primary',
    'elm-color__accent',
    'elm-color__warn',
    'elm-color__alert',
    'elm-color__default--bold',
    'elm-color__primary--bold',
    'elm-color__accent--bold',
    'elm-color__warn--bold',
    'elm-color__alert--bold'
];